# docs/openapi.notifier.json
{"openapi":"3.1.0","info":{"title":"docswarm notifier · WhatsApp/Gmail + firma electrónica","description":"Microservicio de comunicación. Endpoints M2M con X-API-Key.","version":"0.1.0"},"paths":{"/health":{"get":{"tags":["health"],"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/pendientes":{"get":{"tags":["notificaciones"],"summary":"Pendientes","operationId":"pendientes_pendientes_get","parameters":[{"name":"canal","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Canal"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":500,"default":100,"title":"Limit"}},{"name":"X-API-Key","in":"header","required":true,"schema":{"type":"string","title":"X-Api-Key"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/notificaciones/{notif_id}/estado":{"post":{"tags":["notificaciones"],"summary":"Actualizar Estado","operationId":"actualizar_estado_notificaciones__notif_id__estado_post","parameters":[{"name":"notif_id","in":"path","required":true,"schema":{"type":"integer","title":"Notif Id"}},{"name":"X-API-Key","in":"header","required":true,"schema":{"type":"string","title":"X-Api-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EstadoUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/firmas":{"post":{"tags":["firmas"],"summary":"Iniciar Firma","operationId":"iniciar_firma_firmas_post","parameters":[{"name":"X-API-Key","in":"header","required":true,"schema":{"type":"string","title":"X-Api-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IniciarFirma"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/firmas/{firma_id}/evento":{"post":{"tags":["firmas"],"summary":"Evento Firma","operationId":"evento_firma_firmas__firma_id__evento_post","parameters":[{"name":"firma_id","in":"path","required":true,"schema":{"type":"integer","title":"Firma Id"}},{"name":"X-API-Key","in":"header","required":true,"schema":{"type":"string","title":"X-Api-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventoFirma"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/dispatch":{"post":{"tags":["dispatch"],"summary":"Dispatch","description":"Toma pendientes y las envía por el canal correspondiente, marcando estado.\n\nÚtil cuando el envío vive aquí (canales en services/notifier/channels). Si el\nenvío lo hace el repo externo, usa /pendientes + /notificaciones/{id}/estado.","operationId":"dispatch_dispatch_post","parameters":[{"name":"canal","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Canal"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":200,"default":50,"title":"Limit"}},{"name":"X-API-Key","in":"header","required":true,"schema":{"type":"string","title":"X-Api-Key"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"EstadoUpdate":{"properties":{"estado":{"type":"string","title":"Estado"},"referencia_externa":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Referencia Externa"},"error":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error"}},"type":"object","required":["estado"],"title":"EstadoUpdate","description":"Callback de entrega que reporta el microservicio externo."},"EventoFirma":{"properties":{"estado":{"type":"string","title":"Estado"},"evidencia":{"additionalProperties":true,"type":"object","title":"Evidencia","default":{}},"referencia_externa":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Referencia Externa"}},"type":"object","required":["estado"],"title":"EventoFirma","description":"Callback del proceso de firma (lo reporta el microservicio externo)."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"IniciarFirma":{"properties":{"afectado_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Afectado Id"},"contrato_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Contrato Id"},"usuario_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Usuario Id"},"notificacion_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Notificacion Id"},"canal":{"type":"string","title":"Canal","default":"WHATSAPP"}},"type":"object","title":"IniciarFirma","description":"Inicia el proceso de firma electrónica (por WhatsApp/Gmail/Web)."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}